Attempt resource client split
