+++ src/azure-cli/azure/cli/command_modules/resource/custom.py
@@ -20,7 +20,8 @@ from urllib.parse import urlparse, unquote
 
 from azure.mgmt.core.tools import is_valid_resource_id, parse_resource_id
 
-from azure.mgmt.resource.resources.models import GenericResource, DeploymentMode
+from azure.mgmt.resource.resources.models import GenericResource
+from azure.mgmt.resource.deployments.models import DeploymentMode
 
 from azure.cli.core.azclierror import ArgumentUsageError, InvalidArgumentValueError, RequiredArgumentMissingError, ResourceNotFoundError
 from azure.cli.core.parser import IncorrectUsageError
@@ -32,7 +33,7 @@ from azure.cli.core.profiles import ResourceType, get_sdk, get_api_version, AZUR
 
 from azure.cli.command_modules.resource._client_factory import (
     _resource_client_factory, _resource_policy_client_factory, _resource_lock_client_factory,
-    _resource_links_client_factory, _resource_deploymentscripts_client_factory, _resource_deploymentstacks_client_factory, _authorization_management_client, _resource_managedapps_client_factory, _resource_templatespecs_client_factory, _resource_privatelinks_client_factory)
+    _resource_links_client_factory, _resource_deployments_client_factory, _resource_deploymentscripts_client_factory, _resource_deploymentstacks_client_factory, _authorization_management_client, _resource_managedapps_client_factory, _resource_templatespecs_client_factory, _resource_privatelinks_client_factory)
 from azure.cli.command_modules.resource._validators import _parse_lock_id
 from azure.cli.command_modules.resource.parameters import StacksActionOnUnmanage
 
@@ -1867,82 +1868,82 @@ def list_applications(cmd, resource_group_name=None):
 
 
 def list_deployments_at_subscription_scope(cmd, filter_string=None):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.list_at_subscription_scope(filter=filter_string)
 
 
 def list_deployments_at_resource_group(cmd, resource_group_name, filter_string=None):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.list_by_resource_group(resource_group_name, filter=filter_string)
 
 
 def list_deployments_at_management_group(cmd, management_group_id, filter_string=None):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.list_at_management_group_scope(management_group_id, filter=filter_string)
 
 
 def list_deployments_at_tenant_scope(cmd, filter_string=None):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.list_at_tenant_scope(filter=filter_string)
 
 
 def get_deployment_at_subscription_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.get_at_subscription_scope(deployment_name)
 
 
 def get_deployment_at_resource_group(cmd, resource_group_name, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.get(resource_group_name, deployment_name)
 
 
 def get_deployment_at_management_group(cmd, management_group_id, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.get_at_management_group_scope(management_group_id, deployment_name)
 
 
 def get_deployment_at_tenant_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.get_at_tenant_scope(deployment_name)
 
 
 def delete_deployment_at_subscription_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.begin_delete_at_subscription_scope(deployment_name)
 
 
 def delete_deployment_at_resource_group(cmd, resource_group_name, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.begin_delete(resource_group_name, deployment_name)
 
 
 def delete_deployment_at_management_group(cmd, management_group_id, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.begin_delete_at_management_group_scope(management_group_id, deployment_name)
 
 
 def delete_deployment_at_tenant_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.begin_delete_at_tenant_scope(deployment_name)
 
 
 def cancel_deployment_at_subscription_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.cancel_at_subscription_scope(deployment_name)
 
 
 def cancel_deployment_at_resource_group(cmd, resource_group_name, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.cancel(resource_group_name, deployment_name)
 
 
 def cancel_deployment_at_management_group(cmd, management_group_id, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.cancel_at_management_group_scope(management_group_id, deployment_name)
 
 
 def cancel_deployment_at_tenant_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployments.cancel_at_tenant_scope(deployment_name)
 
 
@@ -1970,35 +1971,35 @@ def validate_arm_template(cmd, resource_group_name, template_file=None, template
 
 
 def export_template_at_subscription_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     result = rcf.deployments.export_template_at_subscription_scope(deployment_name)
 
     print(json.dumps(result.template, indent=2))  # pylint: disable=no-member
 
 
 def export_template_at_resource_group(cmd, resource_group_name, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     result = rcf.deployments.export_template(resource_group_name, deployment_name)
 
     print(json.dumps(result.template, indent=2))  # pylint: disable=no-member
 
 
 def export_template_at_management_group(cmd, management_group_id, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     result = rcf.deployments.export_template_at_management_group_scope(management_group_id, deployment_name)
 
     print(json.dumps(result.template, indent=2))  # pylint: disable=no-member
 
 
 def export_template_at_tenant_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     result = rcf.deployments.export_template_at_tenant_scope(deployment_name)
 
     print(json.dumps(result.template, indent=2))  # pylint: disable=no-member
 
 
 def export_deployment_as_template(cmd, resource_group_name, deployment_name):
-    smc = _resource_client_factory(cmd.cli_ctx)
+    smc = _resource_deployments_client_factory(cmd.cli_ctx)
     result = smc.deployments.export_template(resource_group_name, deployment_name)
     print(json.dumps(result.template, indent=2))  # pylint: disable=no-member
 
@@ -3042,42 +3043,42 @@ def export_template_deployment_stack_at_management_group(cmd, management_group_i
 
 
 def list_deployment_operations_at_subscription_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.list_at_subscription_scope(deployment_name)
 
 
 def list_deployment_operations_at_resource_group(cmd, resource_group_name, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.list(resource_group_name, deployment_name)
 
 
 def list_deployment_operations_at_management_group(cmd, management_group_id, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.list_at_management_group_scope(management_group_id, deployment_name)
 
 
 def list_deployment_operations_at_tenant_scope(cmd, deployment_name):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.list_at_tenant_scope(deployment_name)
 
 
 def get_deployment_operation_at_subscription_scope(cmd, deployment_name, op_id):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.get_at_subscription_scope(deployment_name, op_id)
 
 
 def get_deployment_operation_at_resource_group(cmd, resource_group_name, deployment_name, op_id):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.get(resource_group_name, deployment_name, op_id)
 
 
 def get_deployment_operation_at_management_group(cmd, management_group_id, deployment_name, op_id):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.get_at_management_group_scope(management_group_id, deployment_name, op_id)
 
 
 def get_deployment_operation_at_tenant_scope(cmd, deployment_name, op_id):
-    rcf = _resource_client_factory(cmd.cli_ctx)
+    rcf = _resource_deployments_client_factory(cmd.cli_ctx)
     return rcf.deployment_operations.get_at_tenant_scope(deployment_name, op_id)
