+++ src/azure-cli/requirements.py3.windows.txt
@@ -66,7 +66,11 @@ azure-mgmt-recoveryservices==3.1.0
 azure-mgmt-recoveryservicesbackup==9.2.0
 azure-mgmt-redhatopenshift==1.5.0
 azure-mgmt-redis==14.5.0
-azure-mgmt-resource==23.3.0
+/workspaces/azure-cli/azure_mgmt_resource-25.0.0-py3-none-any.whl
+azure-mgmt-resource-deployments==1.0.0b1
+azure-mgmt-resource-deploymentscripts==1.0.0b1
+azure-mgmt-resource-deploymentstacks==1.0.0b1
+azure-mgmt-resource-templatespecs==1.0.0b1
 azure-mgmt-search==9.0.0
 azure-mgmt-security==6.0.0
 azure-mgmt-servicebus==8.2.0
