+++ src/azure-cli/setup.py
@@ -109,7 +109,11 @@ DEPENDENCIES = [
     'azure-mgmt-recoveryservices~=3.1.0',
     'azure-mgmt-redhatopenshift~=1.5.0',
     'azure-mgmt-redis~=14.5.0',
-    'azure-mgmt-resource==23.3.0',
+    # TODO add resource package here
+    'azure-mgmt-resource-deployments==1.0.0b1',
+    'azure-mgmt-resource-deploymentscripts==1.0.0b1',
+    'azure-mgmt-resource-deploymentstacks==1.0.0b1',
+    'azure-mgmt-resource-templatespecs==1.0.0b1',
     'azure-mgmt-search~=9.0',
     'azure-mgmt-security==6.0.0',
     'azure-mgmt-servicebus~=8.2.0',
